Circle fixes: consolidate Obj Inspector Items into one file
- Moved all individual items into `index.tsx`
- Moved all util functions into `utils.ts`

This eliminates a few more cycles
